Fix use of a pinpad reader CKF_PROTECTED_AUTHENTICATION_PATH
When a pinpad reader is used then NULL can be used a PIN value to tell the
PKCS#11 libarry to use the pinpad to enter the PIN.

Some PKCS#11 libraries (like Gemalto IDGO 800) also check the PIN length.
If pPIN is NULL but ulPinLen is not 0 then CKR_ARGUMENTS_BAD is returned.

The new code now forces the PIN length to 0 for a NULL pin.