Makefile: swig do not have a -py3 argument any more
SWIG version 3.0.12 do not document a "-py3" argument any more.
So remove this undocumented option.

The .cpp file genarated with or without -py3 are the same.
The .py file genarated with or without -py3 are the same except that the
-py3 version has Python annotation added:
--- PyKCS11/LowLevel.py2    2017-02-15 10:38:41.000000000 +0100
+++ PyKCS11/LowLevel.py3    2017-02-15 10:38:41.000000000 +0100
@@ -96,11 +96,11 @@
     _newclass = 0

-def cdata(ptr, nelements=1):
+def cdata(ptr: 'void *', nelements: 'size_t'=1) -> "SWIGCDATA":
     return _LowLevel.cdata(ptr, nelements)
 cdata = _LowLevel.cdata

-def memmove(data, indata):
+def memmove(data: 'void *', indata: 'void const *') -> "void":
     return _LowLevel.memmove(data, indata)
 memmove = _LowLevel.memmove
 class SwigPyIterator(_object):

See https://github.com/LudovicRousseau/PyKCS11/pull/3 for more details.