ckbytelist: remove possibility to give a initial size
ckbytelist() now only accept a list of values or None as the initial value.

The problem was that if you use a very big integer as the init parameter then
ckbytelist() would try to allocate a very large memory and would fail.
Integers shall be converted to a list of bytes by the application.

Thanks to Gonçalo Grilo for the bug report
" Session.sign will fill the memory and cause a SIGABRT when data argument is a very big integer #26 "
https://github.com/LudovicRousseau/PyKCS11/issues/26