samples LowLevel: use Python 3
Change the #! to use #!/usr/bin/env python3

Fixes a Debian lintian warning:
P: python3-pykcs11: example-unusual-interpreter python [usr/share/doc/python3-pykcs11/examples/LowLevel/InitTokenPin.py]
N:
N:   This package contains an example script for an interpreter that is not
N:   shipped in the package and is not known to Lintian. It is possible that
N:   there is a typo or the interpreter is not executable.
N:
N:   Visibility: pedantic
N:   Show-Always: no
N:   Check: scripts