Update load()/__del__() to handle Python 3.12 garbage collection
Python 3.12 can perform the garbage collection at any time, not only
just after the object is destroyed.

The problem is that if the PKCS#11 library is loaded twice (or more) a
call to C_Finalize() in one instance will invalidate the calls to the
other instance.

The code used m_bFinalizeOnClose and m_bAutoInitialized to try to avoid
the problem but that was not enough.
We now have a counter of library user and call C_Finalize() only when
the last user release the object.

See test/test_gc.py for a sample case.