Fix reference counting in PyKCS11Lib.load()
Fixes incorrect reference counting for PyKCS11Lib._loaded_libs

The bug can be reproduced with the following steps (this case is
implemented in test_specific_load):

- Create an instance of PyKCS11Lib and load a PKCS#11 library
- Load another library using PyKCS11Lib.load method. At this step
  PyKCS11Lib doesn't unload the first library and decrement the
  reference counter on it, i.e. we have two entries in _loaded_libs map
  pointing to the same CPKCS11Lib object.
- Unload the second library: PyKCS11Lib removes the corresponding entry
  from _loaded_libs and invokes self.lib.Unload(), so now we have a
  dangling entry of the first library pointing to the same CPKCS11Lib
  object, on which Unload has been called.
- Load the first library again: PyKCS11Lib calls Duplicate with the
  dangling entry in _loaded_libs.
- Now, try to perform any operation with PyKCS11Lib => causes a segfault

While investigating this problem I found it convenient to add a
PyKCS11Lib.unload method, which allows to explicitly unload the PKCS#11
library. I also added a couple of tests which require two different
PKCS#11 libraries to run, which is not the case in the testing
enviroment (with a single SoftHSM) you are currently using, so I had to
create a temporary copy of libsofthsm2.so to make them work.

By the way, the current implementation of load/unload is not thread
safe, since it operates with an unprotected shared map (_loaded_libs).
However, I'm not sure how much of an issue this is. I'd say, just
marking these methods as non thread safe is pretty much acceptable as
for me.