bugfix: store CKM_CONCATENATE_BASE_AND_KEY parameter in mechanism context
It's needed to be sure that encKey will not be destroyed by the garbage
collector while the mechanism object is still in use. I guess this is
necessary because the assignment:

self._mech.pParameter = self._encKey

internally copies only a pointer to self._encKey, which would get
invalidated if one destroys self._encKey and that's why it should be
attached to self. It wasn't obvious for me either, but this is how it
works (I guess). The other mechanisms do pretty much the same thing --
i.e. they store the parameter in self._param usually.

Note:
.pParameter is just a property of a CK_MECHANISM() object.