Update OpenSC headers to comply PKCS#11 3.0 specification
- Update PKCS#11 interface version to 3.0.
- Update the definitions of CK_GCM_PARAMS, CK_RSA_PKCS_OAEP_PARAMS,
  CK_ECDH1_DERIVE_PARAMS in accordance with PKCS#11 3.0 headers.

The latest pkcs11.h header from OpenSC was used as the reference.

The reason I need the update is to be able to seamlessly build the
project with the official set of PKCS#11 3.0 headers provided by Oasis.
In fact I need to use a custom version of PyKCS11 with some vendor
specific definitions added in the headers.