setup: Rename the project in lower case
This email is notifying you of an upcoming deprecation that we have
determined may affect you as a result of your recent upload to
'PyKCS11'.

In the future, PyPI will require all newly uploaded binary distribution
filenames to comply with the binary distribution format. Any binary
distributions already uploaded will remain in place as-is and do not
need to be updated.

Specifically, your recent upload of
'PyKCS11-1.5.18-cp310-cp310-win_amd64.whl' is incompatible with the
distribution format specification because the filename does not contain
the normalized project name 'pykcs11'.